+++ src/KingmakerGunslinger/Misfires/FirearmMisfireConditionTransition.cs
@@ -9,6 +9,7 @@ namespace KingmakerGunslinger.Misfires
         None = 0,
         NormalToBroken = 1,
         BrokenToWrecked = 2,
-        AdvancedBrokenRemainsBroken = 3
+        AdvancedBrokenRemainsBroken = 3,
+        ExpertLoadingBrokenRemainsBroken = 4
     }
 }